// packages/iteam-se/typings/iteamse.d.ts
export interface CasesPageQuery {
  pageCases: {
    __typename: 'PageCases'
    headerImage: string | null
    headerText1: string | null
    headerText2: string | null
    headerTextBgColor: string | null
    contactTitle: string
    team: Array<string> | null
  }
  cases: Array<{
    __typename: 'Case'
    location: string
    shortDescription: string
    slug: string
    thumbnailImage: string | null
    title: string
    tags: Array<string | null>
  } | null>
}